move geohash to types
